Move tests to more appropriate locations
Since our tests in the crate root were written, some of the production
code they test has moved to other modules. This commit moves the tests
to those modules.
