+++ src/lib.rs
@@ -5872,8 +5872,6 @@ pub use alloc_support::*;
 #[cfg(test)]
 #[allow(clippy::assertions_on_result_states, clippy::unreadable_literal)]
 mod tests {
-    use core::convert::TryInto as _;
-
     use static_assertions::assert_impl_all;
 
     use super::*;
@@ -5898,815 +5896,6 @@ mod tests {
             // [1] https://github.com/rust-lang/unsafe-code-guidelines/issues/375
             unsafe { mem::transmute(slc) }
         }
-    }
-
-    /// Tests of when a sized `DstLayout` is extended with a sized field.
-    #[allow(clippy::decimal_literal_representation)]
-    #[test]
-    fn test_dst_layout_extend_sized_with_sized() {
-        // This macro constructs a layout corresponding to a `u8` and extends it
-        // with a zero-sized trailing field of given alignment `n`. The macro
-        // tests that the resulting layout has both size and alignment `min(n,
-        // P)` for all valid values of `repr(packed(P))`.
-        macro_rules! test_align_is_size {
-            ($n:expr) => {
-                let base = DstLayout::for_type::<u8>();
-                let trailing_field = DstLayout::for_type::<elain::Align<$n>>();
-
-                let packs =
-                    core::iter::once(None).chain((0..29).map(|p| NonZeroUsize::new(2usize.pow(p))));
-
-                for pack in packs {
-                    let composite = base.extend(trailing_field, pack);
-                    let max_align = pack.unwrap_or(DstLayout::CURRENT_MAX_ALIGN);
-                    let align = $n.min(max_align.get());
-                    assert_eq!(
-                        composite,
-                        DstLayout {
-                            align: NonZeroUsize::new(align).unwrap(),
-                            size_info: SizeInfo::Sized { size: align }
-                        }
-                    )
-                }
-            };
-        }
-
-        test_align_is_size!(1);
-        test_align_is_size!(2);
-        test_align_is_size!(4);
-        test_align_is_size!(8);
-        test_align_is_size!(16);
-        test_align_is_size!(32);
-        test_align_is_size!(64);
-        test_align_is_size!(128);
-        test_align_is_size!(256);
-        test_align_is_size!(512);
-        test_align_is_size!(1024);
-        test_align_is_size!(2048);
-        test_align_is_size!(4096);
-        test_align_is_size!(8192);
-        test_align_is_size!(16384);
-        test_align_is_size!(32768);
-        test_align_is_size!(65536);
-        test_align_is_size!(131072);
-        test_align_is_size!(262144);
-        test_align_is_size!(524288);
-        test_align_is_size!(1048576);
-        test_align_is_size!(2097152);
-        test_align_is_size!(4194304);
-        test_align_is_size!(8388608);
-        test_align_is_size!(16777216);
-        test_align_is_size!(33554432);
-        test_align_is_size!(67108864);
-        test_align_is_size!(33554432);
-        test_align_is_size!(134217728);
-        test_align_is_size!(268435456);
-    }
-
-    /// Tests of when a sized `DstLayout` is extended with a DST field.
-    #[test]
-    fn test_dst_layout_extend_sized_with_dst() {
-        // Test that for all combinations of real-world alignments and
-        // `repr_packed` values, that the extension of a sized `DstLayout`` with
-        // a DST field correctly computes the trailing offset in the composite
-        // layout.
-
-        let aligns = (0..29).map(|p| NonZeroUsize::new(2usize.pow(p)).unwrap());
-        let packs = core::iter::once(None).chain(aligns.clone().map(Some));
-
-        for align in aligns {
-            for pack in packs.clone() {
-                let base = DstLayout::for_type::<u8>();
-                let elem_size = 42;
-                let trailing_field_offset = 11;
-
-                let trailing_field = DstLayout {
-                    align,
-                    size_info: SizeInfo::SliceDst(TrailingSliceLayout { elem_size, offset: 11 }),
-                };
-
-                let composite = base.extend(trailing_field, pack);
-
-                let max_align = pack.unwrap_or(DstLayout::CURRENT_MAX_ALIGN).get();
-
-                let align = align.get().min(max_align);
-
-                assert_eq!(
-                    composite,
-                    DstLayout {
-                        align: NonZeroUsize::new(align).unwrap(),
-                        size_info: SizeInfo::SliceDst(TrailingSliceLayout {
-                            elem_size,
-                            offset: align + trailing_field_offset,
-                        }),
-                    }
-                )
-            }
-        }
-    }
-
-    /// Tests that calling `pad_to_align` on a sized `DstLayout` adds the
-    /// expected amount of trailing padding.
-    #[test]
-    fn test_dst_layout_pad_to_align_with_sized() {
-        // For all valid alignments `align`, construct a one-byte layout aligned
-        // to `align`, call `pad_to_align`, and assert that the size of the
-        // resulting layout is equal to `align`.
-        for align in (0..29).map(|p| NonZeroUsize::new(2usize.pow(p)).unwrap()) {
-            let layout = DstLayout { align, size_info: SizeInfo::Sized { size: 1 } };
-
-            assert_eq!(
-                layout.pad_to_align(),
-                DstLayout { align, size_info: SizeInfo::Sized { size: align.get() } }
-            );
-        }
-
-        // Test explicitly-provided combinations of unpadded and padded
-        // counterparts.
-
-        macro_rules! test {
-            (unpadded { size: $unpadded_size:expr, align: $unpadded_align:expr }
-                => padded { size: $padded_size:expr, align: $padded_align:expr }) => {
-                let unpadded = DstLayout {
-                    align: NonZeroUsize::new($unpadded_align).unwrap(),
-                    size_info: SizeInfo::Sized { size: $unpadded_size },
-                };
-                let padded = unpadded.pad_to_align();
-
-                assert_eq!(
-                    padded,
-                    DstLayout {
-                        align: NonZeroUsize::new($padded_align).unwrap(),
-                        size_info: SizeInfo::Sized { size: $padded_size },
-                    }
-                );
-            };
-        }
-
-        test!(unpadded { size: 0, align: 4 } => padded { size: 0, align: 4 });
-        test!(unpadded { size: 1, align: 4 } => padded { size: 4, align: 4 });
-        test!(unpadded { size: 2, align: 4 } => padded { size: 4, align: 4 });
-        test!(unpadded { size: 3, align: 4 } => padded { size: 4, align: 4 });
-        test!(unpadded { size: 4, align: 4 } => padded { size: 4, align: 4 });
-        test!(unpadded { size: 5, align: 4 } => padded { size: 8, align: 4 });
-        test!(unpadded { size: 6, align: 4 } => padded { size: 8, align: 4 });
-        test!(unpadded { size: 7, align: 4 } => padded { size: 8, align: 4 });
-        test!(unpadded { size: 8, align: 4 } => padded { size: 8, align: 4 });
-
-        let current_max_align = DstLayout::CURRENT_MAX_ALIGN.get();
-
-        test!(unpadded { size: 1, align: current_max_align }
-            => padded { size: current_max_align, align: current_max_align });
-
-        test!(unpadded { size: current_max_align + 1, align: current_max_align }
-            => padded { size: current_max_align * 2, align: current_max_align });
-    }
-
-    /// Tests that calling `pad_to_align` on a DST `DstLayout` is a no-op.
-    #[test]
-    fn test_dst_layout_pad_to_align_with_dst() {
-        for align in (0..29).map(|p| NonZeroUsize::new(2usize.pow(p)).unwrap()) {
-            for offset in 0..10 {
-                for elem_size in 0..10 {
-                    let layout = DstLayout {
-                        align,
-                        size_info: SizeInfo::SliceDst(TrailingSliceLayout { offset, elem_size }),
-                    };
-                    assert_eq!(layout.pad_to_align(), layout);
-                }
-            }
-        }
-    }
-
-    // This test takes a long time when running under Miri, so we skip it in
-    // that case. This is acceptable because this is a logic test that doesn't
-    // attempt to expose UB.
-    #[test]
-    #[cfg_attr(miri, ignore)]
-    fn test_validate_cast_and_convert_metadata() {
-        #[allow(non_local_definitions)]
-        impl From<usize> for SizeInfo {
-            fn from(size: usize) -> SizeInfo {
-                SizeInfo::Sized { size }
-            }
-        }
-
-        #[allow(non_local_definitions)]
-        impl From<(usize, usize)> for SizeInfo {
-            fn from((offset, elem_size): (usize, usize)) -> SizeInfo {
-                SizeInfo::SliceDst(TrailingSliceLayout { offset, elem_size })
-            }
-        }
-
-        fn layout<S: Into<SizeInfo>>(s: S, align: usize) -> DstLayout {
-            DstLayout { size_info: s.into(), align: NonZeroUsize::new(align).unwrap() }
-        }
-
-        /// This macro accepts arguments in the form of:
-        ///
-        ///           layout(_, _, _).validate(_, _, _), Ok(Some((_, _)))
-        ///                  |  |  |           |  |  |            |  |
-        ///    base_size ----+  |  |           |  |  |            |  |
-        ///    align -----------+  |           |  |  |            |  |
-        ///    trailing_size ------+           |  |  |            |  |
-        ///    addr ---------------------------+  |  |            |  |
-        ///    bytes_len -------------------------+  |            |  |
-        ///    cast_type ----------------------------+            |  |
-        ///    elems ---------------------------------------------+  |
-        ///    split_at ---------------------------------------------+
-        ///
-        /// `.validate` is shorthand for `.validate_cast_and_convert_metadata`
-        /// for brevity.
-        ///
-        /// Each argument can either be an iterator or a wildcard. Each
-        /// wildcarded variable is implicitly replaced by an iterator over a
-        /// representative sample of values for that variable. Each `test!`
-        /// invocation iterates over every combination of values provided by
-        /// each variable's iterator (ie, the cartesian product) and validates
-        /// that the results are expected.
-        ///
-        /// The final argument uses the same syntax, but it has a different
-        /// meaning:
-        /// - If it is `Ok(pat)`, then the pattern `pat` is supplied to
-        ///   `assert_matches!` to validate the computed result for each
-        ///   combination of input values.
-        /// - If it is `Err(Some(msg) | None)`, then `test!` validates that the
-        ///   call to `validate_cast_and_convert_metadata` panics with the given
-        ///   panic message or, if the current Rust toolchain version is too
-        ///   early to support panicking in `const fn`s, panics with *some*
-        ///   message. In the latter case, the `const_panic!` macro is used,
-        ///   which emits code which causes a non-panicking error at const eval
-        ///   time, but which does panic when invoked at runtime. Thus, it is
-        ///   merely difficult to predict the *value* of this panic. We deem
-        ///   that testing against the real panic strings on stable and nightly
-        ///   toolchains is enough to ensure correctness.
-        ///
-        /// Note that the meta-variables that match these variables have the
-        /// `tt` type, and some valid expressions are not valid `tt`s (such as
-        /// `a..b`). In this case, wrap the expression in parentheses, and it
-        /// will become valid `tt`.
-        macro_rules! test {
-            ($(:$sizes:expr =>)?
-                layout($size:tt, $align:tt)
-                .validate($addr:tt, $bytes_len:tt, $cast_type:tt), $expect:pat $(,)?
-            ) => {
-                itertools::iproduct!(
-                    test!(@generate_size $size),
-                    test!(@generate_align $align),
-                    test!(@generate_usize $addr),
-                    test!(@generate_usize $bytes_len),
-                    test!(@generate_cast_type $cast_type)
-                ).for_each(|(size_info, align, addr, bytes_len, cast_type)| {
-                    // Temporarily disable the panic hook installed by the test
-                    // harness. If we don't do this, all panic messages will be
-                    // kept in an internal log. On its own, this isn't a
-                    // problem, but if a non-caught panic ever happens (ie, in
-                    // code later in this test not in this macro), all of the
-                    // previously-buffered messages will be dumped, hiding the
-                    // real culprit.
-                    let previous_hook = std::panic::take_hook();
-                    // I don't understand why, but this seems to be required in
-                    // addition to the previous line.
-                    std::panic::set_hook(Box::new(|_| {}));
-                    let actual = std::panic::catch_unwind(|| {
-                        layout(size_info, align).validate_cast_and_convert_metadata(addr, bytes_len, cast_type)
-                    }).map_err(|d| {
-                        let msg = d.downcast::<&'static str>().ok().map(|s| *s.as_ref());
-                        assert!(msg.is_some() || cfg!(not(zerocopy_panic_in_const)), "non-string panic messages are not permitted when `--cfg zerocopy_panic_in_const` is set");
-                        msg
-                    });
-                    std::panic::set_hook(previous_hook);
-
-                    assert_matches::assert_matches!(
-                        actual, $expect,
-                        "layout({:?}, {}).validate_cast_and_convert_metadata({}, {}, {:?})" ,size_info, align, addr, bytes_len, cast_type
-                    );
-                });
-            };
-            (@generate_usize _) => { 0..8 };
-            // Generate sizes for both Sized and !Sized types.
-            (@generate_size _) => {
-                test!(@generate_size (_)).chain(test!(@generate_size (_, _)))
-            };
-            // Generate sizes for both Sized and !Sized types by chaining
-            // specified iterators for each.
-            (@generate_size ($sized_sizes:tt | $unsized_sizes:tt)) => {
-                test!(@generate_size ($sized_sizes)).chain(test!(@generate_size $unsized_sizes))
-            };
-            // Generate sizes for Sized types.
-            (@generate_size (_)) => { test!(@generate_size (0..8)) };
-            (@generate_size ($sizes:expr)) => { $sizes.into_iter().map(Into::<SizeInfo>::into) };
-            // Generate sizes for !Sized types.
-            (@generate_size ($min_sizes:tt, $elem_sizes:tt)) => {
-                itertools::iproduct!(
-                    test!(@generate_min_size $min_sizes),
-                    test!(@generate_elem_size $elem_sizes)
-                ).map(Into::<SizeInfo>::into)
-            };
-            (@generate_fixed_size _) => { (0..8).into_iter().map(Into::<SizeInfo>::into) };
-            (@generate_min_size _) => { 0..8 };
-            (@generate_elem_size _) => { 1..8 };
-            (@generate_align _) => { [1, 2, 4, 8, 16] };
-            (@generate_opt_usize _) => { [None].into_iter().chain((0..8).map(Some).into_iter()) };
-            (@generate_cast_type _) => { [CastType::Prefix, CastType::Suffix] };
-            (@generate_cast_type $variant:ident) => { [CastType::$variant] };
-            // Some expressions need to be wrapped in parentheses in order to be
-            // valid `tt`s (required by the top match pattern). See the comment
-            // below for more details. This arm removes these parentheses to
-            // avoid generating an `unused_parens` warning.
-            (@$_:ident ($vals:expr)) => { $vals };
-            (@$_:ident $vals:expr) => { $vals };
-        }
-
-        const EVENS: [usize; 8] = [0, 2, 4, 6, 8, 10, 12, 14];
-        const ODDS: [usize; 8] = [1, 3, 5, 7, 9, 11, 13, 15];
-
-        // base_size is too big for the memory region.
-        test!(
-            layout(((1..8) | ((1..8), (1..8))), _).validate([0], [0], _),
-            Ok(Err(MetadataCastError::Size))
-        );
-        test!(
-            layout(((2..8) | ((2..8), (2..8))), _).validate([0], [1], Prefix),
-            Ok(Err(MetadataCastError::Size))
-        );
-        test!(
-            layout(((2..8) | ((2..8), (2..8))), _).validate([0x1000_0000 - 1], [1], Suffix),
-            Ok(Err(MetadataCastError::Size))
-        );
-
-        // addr is unaligned for prefix cast
-        test!(layout(_, [2]).validate(ODDS, _, Prefix), Ok(Err(MetadataCastError::Alignment)));
-        test!(layout(_, [2]).validate(ODDS, _, Prefix), Ok(Err(MetadataCastError::Alignment)));
-
-        // addr is aligned, but end of buffer is unaligned for suffix cast
-        test!(layout(_, [2]).validate(EVENS, ODDS, Suffix), Ok(Err(MetadataCastError::Alignment)));
-        test!(layout(_, [2]).validate(EVENS, ODDS, Suffix), Ok(Err(MetadataCastError::Alignment)));
-
-        // Unfortunately, these constants cannot easily be used in the
-        // implementation of `validate_cast_and_convert_metadata`, since
-        // `panic!` consumes a string literal, not an expression.
-        //
-        // It's important that these messages be in a separate module. If they
-        // were at the function's top level, we'd pass them to `test!` as, e.g.,
-        // `Err(TRAILING)`, which would run into a subtle Rust footgun - the
-        // `TRAILING` identifier would be treated as a pattern to match rather
-        // than a value to check for equality.
-        mod msgs {
-            pub(super) const TRAILING: &str =
-                "attempted to cast to slice type with zero-sized element";
-            pub(super) const OVERFLOW: &str = "`addr` + `bytes_len` > usize::MAX";
-        }
-
-        // casts with ZST trailing element types are unsupported
-        test!(layout((_, [0]), _).validate(_, _, _), Err(Some(msgs::TRAILING) | None),);
-
-        // addr + bytes_len must not overflow usize
-        test!(layout(_, _).validate([usize::MAX], (1..100), _), Err(Some(msgs::OVERFLOW) | None));
-        test!(layout(_, _).validate((1..100), [usize::MAX], _), Err(Some(msgs::OVERFLOW) | None));
-        test!(
-            layout(_, _).validate(
-                [usize::MAX / 2 + 1, usize::MAX],
-                [usize::MAX / 2 + 1, usize::MAX],
-                _
-            ),
-            Err(Some(msgs::OVERFLOW) | None)
-        );
-
-        // Validates that `validate_cast_and_convert_metadata` satisfies its own
-        // documented safety postconditions, and also a few other properties
-        // that aren't documented but we want to guarantee anyway.
-        fn validate_behavior(
-            (layout, addr, bytes_len, cast_type): (DstLayout, usize, usize, CastType),
-        ) {
-            if let Ok((elems, split_at)) =
-                layout.validate_cast_and_convert_metadata(addr, bytes_len, cast_type)
-            {
-                let (size_info, align) = (layout.size_info, layout.align);
-                let debug_str = format!(
-                    "layout({:?}, {}).validate_cast_and_convert_metadata({}, {}, {:?}) => ({}, {})",
-                    size_info, align, addr, bytes_len, cast_type, elems, split_at
-                );
-
-                // If this is a sized type (no trailing slice), then `elems` is
-                // meaningless, but in practice we set it to 0. Callers are not
-                // allowed to rely on this, but a lot of math is nicer if
-                // they're able to, and some callers might accidentally do that.
-                let sized = matches!(layout.size_info, SizeInfo::Sized { .. });
-                assert!(!(sized && elems != 0), "{}", debug_str);
-
-                let resulting_size = match layout.size_info {
-                    SizeInfo::Sized { size } => size,
-                    SizeInfo::SliceDst(TrailingSliceLayout { offset, elem_size }) => {
-                        let padded_size = |elems| {
-                            let without_padding = offset + elems * elem_size;
-                            without_padding + util::padding_needed_for(without_padding, align)
-                        };
-
-                        let resulting_size = padded_size(elems);
-                        // Test that `validate_cast_and_convert_metadata`
-                        // computed the largest possible value that fits in the
-                        // given range.
-                        assert!(padded_size(elems + 1) > bytes_len, "{}", debug_str);
-                        resulting_size
-                    }
-                };
-
-                // Test safety postconditions guaranteed by
-                // `validate_cast_and_convert_metadata`.
-                assert!(resulting_size <= bytes_len, "{}", debug_str);
-                match cast_type {
-                    CastType::Prefix => {
-                        assert_eq!(addr % align, 0, "{}", debug_str);
-                        assert_eq!(resulting_size, split_at, "{}", debug_str);
-                    }
-                    CastType::Suffix => {
-                        assert_eq!(split_at, bytes_len - resulting_size, "{}", debug_str);
-                        assert_eq!((addr + split_at) % align, 0, "{}", debug_str);
-                    }
-                }
-            } else {
-                let min_size = match layout.size_info {
-                    SizeInfo::Sized { size } => size,
-                    SizeInfo::SliceDst(TrailingSliceLayout { offset, .. }) => {
-                        offset + util::padding_needed_for(offset, layout.align)
-                    }
-                };
-
-                // If a cast is invalid, it is either because...
-                // 1. there are insufficent bytes at the given region for type:
-                let insufficient_bytes = bytes_len < min_size;
-                // 2. performing the cast would misalign type:
-                let base = match cast_type {
-                    CastType::Prefix => 0,
-                    CastType::Suffix => bytes_len,
-                };
-                let misaligned = (base + addr) % layout.align != 0;
-
-                assert!(insufficient_bytes || misaligned);
-            }
-        }
-
-        let sizes = 0..8;
-        let elem_sizes = 1..8;
-        let size_infos = sizes
-            .clone()
-            .map(Into::<SizeInfo>::into)
-            .chain(itertools::iproduct!(sizes, elem_sizes).map(Into::<SizeInfo>::into));
-        let layouts = itertools::iproduct!(size_infos, [1, 2, 4, 8, 16, 32])
-            .filter(|(size_info, align)| !matches!(size_info, SizeInfo::Sized { size } if size % align != 0))
-            .map(|(size_info, align)| layout(size_info, align));
-        itertools::iproduct!(layouts, 0..8, 0..8, [CastType::Prefix, CastType::Suffix])
-            .for_each(validate_behavior);
-    }
-
-    #[test]
-    #[cfg(__INTERNAL_USE_ONLY_NIGHTLY_FEATURES_IN_TESTS)]
-    fn test_validate_rust_layout() {
-        use core::ptr::NonNull;
-
-        // This test synthesizes pointers with various metadata and uses Rust's
-        // built-in APIs to confirm that Rust makes decisions about type layout
-        // which are consistent with what we believe is guaranteed by the
-        // language. If this test fails, it doesn't just mean our code is wrong
-        // - it means we're misunderstanding the language's guarantees.
-
-        #[derive(Debug)]
-        struct MacroArgs {
-            offset: usize,
-            align: NonZeroUsize,
-            elem_size: Option<usize>,
-        }
-
-        /// # Safety
-        ///
-        /// `test` promises to only call `addr_of_slice_field` on a `NonNull<T>`
-        /// which points to a valid `T`.
-        ///
-        /// `with_elems` must produce a pointer which points to a valid `T`.
-        fn test<T: ?Sized, W: Fn(usize) -> NonNull<T>>(
-            args: MacroArgs,
-            with_elems: W,
-            addr_of_slice_field: Option<fn(NonNull<T>) -> NonNull<u8>>,
-        ) {
-            let dst = args.elem_size.is_some();
-            let layout = {
-                let size_info = match args.elem_size {
-                    Some(elem_size) => {
-                        SizeInfo::SliceDst(TrailingSliceLayout { offset: args.offset, elem_size })
-                    }
-                    None => SizeInfo::Sized {
-                        // Rust only supports types whose sizes are a multiple
-                        // of their alignment. If the macro created a type like
-                        // this:
-                        //
-                        //   #[repr(C, align(2))]
-                        //   struct Foo([u8; 1]);
-                        //
-                        // ...then Rust will automatically round the type's size
-                        // up to 2.
-                        size: args.offset + util::padding_needed_for(args.offset, args.align),
-                    },
-                };
-                DstLayout { size_info, align: args.align }
-            };
-
-            for elems in 0..128 {
-                let ptr = with_elems(elems);
-
-                if let Some(addr_of_slice_field) = addr_of_slice_field {
-                    let slc_field_ptr = addr_of_slice_field(ptr).as_ptr();
-                    // SAFETY: Both `slc_field_ptr` and `ptr` are pointers to
-                    // the same valid Rust object.
-                    #[allow(clippy::incompatible_msrv)]
-                    // Work around https://github.com/rust-lang/rust-clippy/issues/12280
-                    let offset: usize =
-                        unsafe { slc_field_ptr.byte_offset_from(ptr.as_ptr()).try_into().unwrap() };
-                    assert_eq!(offset, args.offset);
-                }
-
-                // SAFETY: `ptr` points to a valid `T`.
-                let (size, align) = unsafe {
-                    (mem::size_of_val_raw(ptr.as_ptr()), mem::align_of_val_raw(ptr.as_ptr()))
-                };
-
-                // Avoid expensive allocation when running under Miri.
-                let assert_msg = if !cfg!(miri) {
-                    format!("\n{:?}\nsize:{}, align:{}", args, size, align)
-                } else {
-                    String::new()
-                };
-
-                let without_padding =
-                    args.offset + args.elem_size.map(|elem_size| elems * elem_size).unwrap_or(0);
-                assert!(size >= without_padding, "{}", assert_msg);
-                assert_eq!(align, args.align.get(), "{}", assert_msg);
-
-                // This encodes the most important part of the test: our
-                // understanding of how Rust determines the layout of repr(C)
-                // types. Sized repr(C) types are trivial, but DST types have
-                // some subtlety. Note that:
-                // - For sized types, `without_padding` is just the size of the
-                //   type that we constructed for `Foo`. Since we may have
-                //   requested a larger alignment, `Foo` may actually be larger
-                //   than this, hence `padding_needed_for`.
-                // - For unsized types, `without_padding` is dynamically
-                //   computed from the offset, the element size, and element
-                //   count. We expect that the size of the object should be
-                //   `offset + elem_size * elems` rounded up to the next
-                //   alignment.
-                let expected_size =
-                    without_padding + util::padding_needed_for(without_padding, args.align);
-                assert_eq!(expected_size, size, "{}", assert_msg);
-
-                // For zero-sized element types,
-                // `validate_cast_and_convert_metadata` just panics, so we skip
-                // testing those types.
-                if args.elem_size.map(|elem_size| elem_size > 0).unwrap_or(true) {
-                    let addr = ptr.addr().get();
-                    let (got_elems, got_split_at) = layout
-                        .validate_cast_and_convert_metadata(addr, size, CastType::Prefix)
-                        .unwrap();
-                    // Avoid expensive allocation when running under Miri.
-                    let assert_msg = if !cfg!(miri) {
-                        format!(
-                            "{}\nvalidate_cast_and_convert_metadata({}, {})",
-                            assert_msg, addr, size,
-                        )
-                    } else {
-                        String::new()
-                    };
-                    assert_eq!(got_split_at, size, "{}", assert_msg);
-                    if dst {
-                        assert!(got_elems >= elems, "{}", assert_msg);
-                        if got_elems != elems {
-                            // If `validate_cast_and_convert_metadata`
-                            // returned more elements than `elems`, that
-                            // means that `elems` is not the maximum number
-                            // of elements that can fit in `size` - in other
-                            // words, there is enough padding at the end of
-                            // the value to fit at least one more element.
-                            // If we use this metadata to synthesize a
-                            // pointer, despite having a different element
-                            // count, we still expect it to have the same
-                            // size.
-                            let got_ptr = with_elems(got_elems);
-                            // SAFETY: `got_ptr` is a pointer to a valid `T`.
-                            let size_of_got_ptr = unsafe { mem::size_of_val_raw(got_ptr.as_ptr()) };
-                            assert_eq!(size_of_got_ptr, size, "{}", assert_msg);
-                        }
-                    } else {
-                        // For sized casts, the returned element value is
-                        // technically meaningless, and we don't guarantee any
-                        // particular value. In practice, it's always zero.
-                        assert_eq!(got_elems, 0, "{}", assert_msg)
-                    }
-                }
-            }
-        }
-
-        macro_rules! validate_against_rust {
-            ($offset:literal, $align:literal $(, $elem_size:literal)?) => {{
-                #[repr(C, align($align))]
-                struct Foo([u8; $offset]$(, [[u8; $elem_size]])?);
-
-                let args = MacroArgs {
-                    offset: $offset,
-                    align: $align.try_into().unwrap(),
-                    elem_size: {
-                        #[allow(unused)]
-                        let ret = None::<usize>;
-                        $(let ret = Some($elem_size);)?
-                        ret
-                    }
-                };
-
-                #[repr(C, align($align))]
-                struct FooAlign;
-                // Create an aligned buffer to use in order to synthesize
-                // pointers to `Foo`. We don't ever load values from these
-                // pointers - we just do arithmetic on them - so having a "real"
-                // block of memory as opposed to a validly-aligned-but-dangling
-                // pointer is only necessary to make Miri happy since we run it
-                // with "strict provenance" checking enabled.
-                let aligned_buf = Align::<_, FooAlign>::new([0u8; 1024]);
-                let with_elems = |elems| {
-                    let slc = NonNull::slice_from_raw_parts(NonNull::from(&aligned_buf.t), elems);
-                    #[allow(clippy::as_conversions)]
-                    NonNull::new(slc.as_ptr() as *mut Foo).unwrap()
-                };
-                let addr_of_slice_field = {
-                    #[allow(unused)]
-                    let f = None::<fn(NonNull<Foo>) -> NonNull<u8>>;
-                    $(
-                        // SAFETY: `test` promises to only call `f` with a `ptr`
-                        // to a valid `Foo`.
-                        let f: Option<fn(NonNull<Foo>) -> NonNull<u8>> = Some(|ptr: NonNull<Foo>| unsafe {
-                            NonNull::new(ptr::addr_of_mut!((*ptr.as_ptr()).1)).unwrap().cast::<u8>()
-                        });
-                        let _ = $elem_size;
-                    )?
-                    f
-                };
-
-                test::<Foo, _>(args, with_elems, addr_of_slice_field);
-            }};
-        }
-
-        // Every permutation of:
-        // - offset in [0, 4]
-        // - align in [1, 16]
-        // - elem_size in [0, 4] (plus no elem_size)
-        validate_against_rust!(0, 1);
-        validate_against_rust!(0, 1, 0);
-        validate_against_rust!(0, 1, 1);
-        validate_against_rust!(0, 1, 2);
-        validate_against_rust!(0, 1, 3);
-        validate_against_rust!(0, 1, 4);
-        validate_against_rust!(0, 2);
-        validate_against_rust!(0, 2, 0);
-        validate_against_rust!(0, 2, 1);
-        validate_against_rust!(0, 2, 2);
-        validate_against_rust!(0, 2, 3);
-        validate_against_rust!(0, 2, 4);
-        validate_against_rust!(0, 4);
-        validate_against_rust!(0, 4, 0);
-        validate_against_rust!(0, 4, 1);
-        validate_against_rust!(0, 4, 2);
-        validate_against_rust!(0, 4, 3);
-        validate_against_rust!(0, 4, 4);
-        validate_against_rust!(0, 8);
-        validate_against_rust!(0, 8, 0);
-        validate_against_rust!(0, 8, 1);
-        validate_against_rust!(0, 8, 2);
-        validate_against_rust!(0, 8, 3);
-        validate_against_rust!(0, 8, 4);
-        validate_against_rust!(0, 16);
-        validate_against_rust!(0, 16, 0);
-        validate_against_rust!(0, 16, 1);
-        validate_against_rust!(0, 16, 2);
-        validate_against_rust!(0, 16, 3);
-        validate_against_rust!(0, 16, 4);
-        validate_against_rust!(1, 1);
-        validate_against_rust!(1, 1, 0);
-        validate_against_rust!(1, 1, 1);
-        validate_against_rust!(1, 1, 2);
-        validate_against_rust!(1, 1, 3);
-        validate_against_rust!(1, 1, 4);
-        validate_against_rust!(1, 2);
-        validate_against_rust!(1, 2, 0);
-        validate_against_rust!(1, 2, 1);
-        validate_against_rust!(1, 2, 2);
-        validate_against_rust!(1, 2, 3);
-        validate_against_rust!(1, 2, 4);
-        validate_against_rust!(1, 4);
-        validate_against_rust!(1, 4, 0);
-        validate_against_rust!(1, 4, 1);
-        validate_against_rust!(1, 4, 2);
-        validate_against_rust!(1, 4, 3);
-        validate_against_rust!(1, 4, 4);
-        validate_against_rust!(1, 8);
-        validate_against_rust!(1, 8, 0);
-        validate_against_rust!(1, 8, 1);
-        validate_against_rust!(1, 8, 2);
-        validate_against_rust!(1, 8, 3);
-        validate_against_rust!(1, 8, 4);
-        validate_against_rust!(1, 16);
-        validate_against_rust!(1, 16, 0);
-        validate_against_rust!(1, 16, 1);
-        validate_against_rust!(1, 16, 2);
-        validate_against_rust!(1, 16, 3);
-        validate_against_rust!(1, 16, 4);
-        validate_against_rust!(2, 1);
-        validate_against_rust!(2, 1, 0);
-        validate_against_rust!(2, 1, 1);
-        validate_against_rust!(2, 1, 2);
-        validate_against_rust!(2, 1, 3);
-        validate_against_rust!(2, 1, 4);
-        validate_against_rust!(2, 2);
-        validate_against_rust!(2, 2, 0);
-        validate_against_rust!(2, 2, 1);
-        validate_against_rust!(2, 2, 2);
-        validate_against_rust!(2, 2, 3);
-        validate_against_rust!(2, 2, 4);
-        validate_against_rust!(2, 4);
-        validate_against_rust!(2, 4, 0);
-        validate_against_rust!(2, 4, 1);
-        validate_against_rust!(2, 4, 2);
-        validate_against_rust!(2, 4, 3);
-        validate_against_rust!(2, 4, 4);
-        validate_against_rust!(2, 8);
-        validate_against_rust!(2, 8, 0);
-        validate_against_rust!(2, 8, 1);
-        validate_against_rust!(2, 8, 2);
-        validate_against_rust!(2, 8, 3);
-        validate_against_rust!(2, 8, 4);
-        validate_against_rust!(2, 16);
-        validate_against_rust!(2, 16, 0);
-        validate_against_rust!(2, 16, 1);
-        validate_against_rust!(2, 16, 2);
-        validate_against_rust!(2, 16, 3);
-        validate_against_rust!(2, 16, 4);
-        validate_against_rust!(3, 1);
-        validate_against_rust!(3, 1, 0);
-        validate_against_rust!(3, 1, 1);
-        validate_against_rust!(3, 1, 2);
-        validate_against_rust!(3, 1, 3);
-        validate_against_rust!(3, 1, 4);
-        validate_against_rust!(3, 2);
-        validate_against_rust!(3, 2, 0);
-        validate_against_rust!(3, 2, 1);
-        validate_against_rust!(3, 2, 2);
-        validate_against_rust!(3, 2, 3);
-        validate_against_rust!(3, 2, 4);
-        validate_against_rust!(3, 4);
-        validate_against_rust!(3, 4, 0);
-        validate_against_rust!(3, 4, 1);
-        validate_against_rust!(3, 4, 2);
-        validate_against_rust!(3, 4, 3);
-        validate_against_rust!(3, 4, 4);
-        validate_against_rust!(3, 8);
-        validate_against_rust!(3, 8, 0);
-        validate_against_rust!(3, 8, 1);
-        validate_against_rust!(3, 8, 2);
-        validate_against_rust!(3, 8, 3);
-        validate_against_rust!(3, 8, 4);
-        validate_against_rust!(3, 16);
-        validate_against_rust!(3, 16, 0);
-        validate_against_rust!(3, 16, 1);
-        validate_against_rust!(3, 16, 2);
-        validate_against_rust!(3, 16, 3);
-        validate_against_rust!(3, 16, 4);
-        validate_against_rust!(4, 1);
-        validate_against_rust!(4, 1, 0);
-        validate_against_rust!(4, 1, 1);
-        validate_against_rust!(4, 1, 2);
-        validate_against_rust!(4, 1, 3);
-        validate_against_rust!(4, 1, 4);
-        validate_against_rust!(4, 2);
-        validate_against_rust!(4, 2, 0);
-        validate_against_rust!(4, 2, 1);
-        validate_against_rust!(4, 2, 2);
-        validate_against_rust!(4, 2, 3);
-        validate_against_rust!(4, 2, 4);
-        validate_against_rust!(4, 4);
-        validate_against_rust!(4, 4, 0);
-        validate_against_rust!(4, 4, 1);
-        validate_against_rust!(4, 4, 2);
-        validate_against_rust!(4, 4, 3);
-        validate_against_rust!(4, 4, 4);
-        validate_against_rust!(4, 8);
-        validate_against_rust!(4, 8, 0);
-        validate_against_rust!(4, 8, 1);
-        validate_against_rust!(4, 8, 2);
-        validate_against_rust!(4, 8, 3);
-        validate_against_rust!(4, 8, 4);
-        validate_against_rust!(4, 16);
-        validate_against_rust!(4, 16, 0);
-        validate_against_rust!(4, 16, 1);
-        validate_against_rust!(4, 16, 2);
-        validate_against_rust!(4, 16, 3);
-        validate_against_rust!(4, 16, 4);
     }
 
     #[test]
@@ -7369,311 +6558,6 @@ mod tests {
     }
 
     #[test]
-    fn test_address() {
-        // Test that the `Deref` and `DerefMut` implementations return a
-        // reference which points to the right region of memory.
-
-        let buf = [0];
-        let r = Ref::<_, u8>::from(&buf[..]).unwrap();
-        let buf_ptr = buf.as_ptr();
-        let deref_ptr: *const u8 = r.deref();
-        assert_eq!(buf_ptr, deref_ptr);
-
-        let buf = [0];
-        let r = Ref::<_, [u8]>::from(&buf[..]).unwrap();
-        let buf_ptr = buf.as_ptr();
-        let deref_ptr = r.deref().as_ptr();
-        assert_eq!(buf_ptr, deref_ptr);
-    }
-
-    // Verify that values written to a `Ref` are properly shared between the
-    // typed and untyped representations, that reads via `deref` and `read`
-    // behave the same, and that writes via `deref_mut` and `write` behave the
-    // same.
-    fn test_new_helper(mut r: Ref<&mut [u8], AU64>) {
-        // assert that the value starts at 0
-        assert_eq!(*r, AU64(0));
-        assert_eq!(r.read(), AU64(0));
-
-        // Assert that values written to the typed value are reflected in the
-        // byte slice.
-        const VAL1: AU64 = AU64(0xFF00FF00FF00FF00);
-        *r = VAL1;
-        assert_eq!(r.bytes(), &VAL1.to_bytes());
-        *r = AU64(0);
-        r.write(VAL1);
-        assert_eq!(r.bytes(), &VAL1.to_bytes());
-
-        // Assert that values written to the byte slice are reflected in the
-        // typed value.
-        const VAL2: AU64 = AU64(!VAL1.0); // different from `VAL1`
-        r.bytes_mut().copy_from_slice(&VAL2.to_bytes()[..]);
-        assert_eq!(*r, VAL2);
-        assert_eq!(r.read(), VAL2);
-    }
-
-    // Verify that values written to a `Ref` are properly shared between the
-    // typed and untyped representations; pass a value with `typed_len` `AU64`s
-    // backed by an array of `typed_len * 8` bytes.
-    fn test_new_helper_slice(mut r: Ref<&mut [u8], [AU64]>, typed_len: usize) {
-        // Assert that the value starts out zeroed.
-        assert_eq!(&*r, vec![AU64(0); typed_len].as_slice());
-
-        // Check the backing storage is the exact same slice.
-        let untyped_len = typed_len * 8;
-        assert_eq!(r.bytes().len(), untyped_len);
-        assert_eq!(r.bytes().as_ptr(), r.as_ptr().cast::<u8>());
-
-        // Assert that values written to the typed value are reflected in the
-        // byte slice.
-        const VAL1: AU64 = AU64(0xFF00FF00FF00FF00);
-        for typed in &mut *r {
-            *typed = VAL1;
-        }
-        assert_eq!(r.bytes(), VAL1.0.to_ne_bytes().repeat(typed_len).as_slice());
-
-        // Assert that values written to the byte slice are reflected in the
-        // typed value.
-        const VAL2: AU64 = AU64(!VAL1.0); // different from VAL1
-        r.bytes_mut().copy_from_slice(&VAL2.0.to_ne_bytes().repeat(typed_len));
-        assert!(r.iter().copied().all(|x| x == VAL2));
-    }
-
-    // Verify that values written to a `Ref` are properly shared between the
-    // typed and untyped representations, that reads via `deref` and `read`
-    // behave the same, and that writes via `deref_mut` and `write` behave the
-    // same.
-    fn test_new_helper_unaligned(mut r: Ref<&mut [u8], [u8; 8]>) {
-        // assert that the value starts at 0
-        assert_eq!(*r, [0; 8]);
-        assert_eq!(r.read(), [0; 8]);
-
-        // Assert that values written to the typed value are reflected in the
-        // byte slice.
-        const VAL1: [u8; 8] = [0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00];
-        *r = VAL1;
-        assert_eq!(r.bytes(), &VAL1);
-        *r = [0; 8];
-        r.write(VAL1);
-        assert_eq!(r.bytes(), &VAL1);
-
-        // Assert that values written to the byte slice are reflected in the
-        // typed value.
-        const VAL2: [u8; 8] = [0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF]; // different from VAL1
-        r.bytes_mut().copy_from_slice(&VAL2[..]);
-        assert_eq!(*r, VAL2);
-        assert_eq!(r.read(), VAL2);
-    }
-
-    // Verify that values written to a `Ref` are properly shared between the
-    // typed and untyped representations; pass a value with `len` `u8`s backed
-    // by an array of `len` bytes.
-    fn test_new_helper_slice_unaligned(mut r: Ref<&mut [u8], [u8]>, len: usize) {
-        // Assert that the value starts out zeroed.
-        assert_eq!(&*r, vec![0u8; len].as_slice());
-
-        // Check the backing storage is the exact same slice.
-        assert_eq!(r.bytes().len(), len);
-        assert_eq!(r.bytes().as_ptr(), r.as_ptr());
-
-        // Assert that values written to the typed value are reflected in the
-        // byte slice.
-        let mut expected_bytes = [0xFF, 0x00].iter().copied().cycle().take(len).collect::<Vec<_>>();
-        r.copy_from_slice(&expected_bytes);
-        assert_eq!(r.bytes(), expected_bytes.as_slice());
-
-        // Assert that values written to the byte slice are reflected in the
-        // typed value.
-        for byte in &mut expected_bytes {
-            *byte = !*byte; // different from `expected_len`
-        }
-        r.bytes_mut().copy_from_slice(&expected_bytes);
-        assert_eq!(&*r, expected_bytes.as_slice());
-    }
-
-    #[test]
-    fn test_new_aligned_sized() {
-        // Test that a properly-aligned, properly-sized buffer works for new,
-        // new_from_prefix, and new_from_suffix, and that new_from_prefix and
-        // new_from_suffix return empty slices. Test that a properly-aligned
-        // buffer whose length is a multiple of the element size works for
-        // new_slice.
-
-        // A buffer with an alignment of 8.
-        let mut buf = Align::<[u8; 8], AU64>::default();
-        // `buf.t` should be aligned to 8, so this should always succeed.
-        test_new_helper(Ref::<_, AU64>::from(&mut buf.t[..]).unwrap());
-        {
-            // In a block so that `r` and `suffix` don't live too long.
-            buf.set_default();
-            let (r, suffix) = Ref::<_, AU64>::from_prefix(&mut buf.t[..]).unwrap();
-            assert!(suffix.is_empty());
-            test_new_helper(r);
-        }
-        {
-            buf.set_default();
-            let (prefix, r) = Ref::<_, AU64>::from_suffix(&mut buf.t[..]).unwrap();
-            assert!(prefix.is_empty());
-            test_new_helper(r);
-        }
-
-        // A buffer with alignment 8 and length 24. We choose this length very
-        // intentionally: if we instead used length 16, then the prefix and
-        // suffix lengths would be identical. In the past, we used length 16,
-        // which resulted in this test failing to discover the bug uncovered in
-        // #506.
-        let mut buf = Align::<[u8; 24], AU64>::default();
-        // `buf.t` should be aligned to 8 and have a length which is a multiple
-        // of `size_of::<AU64>()`, so this should always succeed.
-        test_new_helper_slice(Ref::<_, [AU64]>::from(&mut buf.t[..]).unwrap(), 3);
-        let ascending: [u8; 24] = (0..24).collect::<Vec<_>>().try_into().unwrap();
-        // 16 ascending bytes followed by 8 zeros.
-        let mut ascending_prefix = ascending;
-        ascending_prefix[16..].copy_from_slice(&[0, 0, 0, 0, 0, 0, 0, 0]);
-        // 8 zeros followed by 16 ascending bytes.
-        let mut ascending_suffix = ascending;
-        ascending_suffix[..8].copy_from_slice(&[0, 0, 0, 0, 0, 0, 0, 0]);
-
-        {
-            buf.t = ascending_suffix;
-            let (r, suffix) = Ref::<_, [AU64]>::from_prefix_with_elems(&mut buf.t[..], 1).unwrap();
-            assert_eq!(suffix, &ascending[8..]);
-            test_new_helper_slice(r, 1);
-        }
-        {
-            buf.t = ascending_prefix;
-            let (prefix, r) = Ref::<_, [AU64]>::from_suffix_with_elems(&mut buf.t[..], 1).unwrap();
-            assert_eq!(prefix, &ascending[..16]);
-            test_new_helper_slice(r, 1);
-        }
-    }
-
-    #[test]
-    fn test_new_unaligned_sized() {
-        // Test that an unaligned, properly-sized buffer works for
-        // `new_unaligned`, `new_unaligned_from_prefix`, and
-        // `new_unaligned_from_suffix`, and that `new_unaligned_from_prefix`
-        // `new_unaligned_from_suffix` return empty slices. Test that an
-        // unaligned buffer whose length is a multiple of the element size works
-        // for `new_slice`.
-
-        let mut buf = [0u8; 8];
-        test_new_helper_unaligned(Ref::<_, [u8; 8]>::unaligned_from(&mut buf[..]).unwrap());
-        {
-            // In a block so that `r` and `suffix` don't live too long.
-            buf = [0u8; 8];
-            let (r, suffix) = Ref::<_, [u8; 8]>::unaligned_from_prefix(&mut buf[..]).unwrap();
-            assert!(suffix.is_empty());
-            test_new_helper_unaligned(r);
-        }
-        {
-            buf = [0u8; 8];
-            let (prefix, r) = Ref::<_, [u8; 8]>::unaligned_from_suffix(&mut buf[..]).unwrap();
-            assert!(prefix.is_empty());
-            test_new_helper_unaligned(r);
-        }
-
-        let mut buf = [0u8; 16];
-        // `buf.t` should be aligned to 8 and have a length which is a multiple
-        // of `size_of::AU64>()`, so this should always succeed.
-        test_new_helper_slice_unaligned(Ref::<_, [u8]>::unaligned_from(&mut buf[..]).unwrap(), 16);
-
-        {
-            buf = [0u8; 16];
-            let (r, suffix) =
-                Ref::<_, [u8]>::unaligned_from_prefix_with_elems(&mut buf[..], 8).unwrap();
-            assert_eq!(suffix, [0; 8]);
-            test_new_helper_slice_unaligned(r, 8);
-        }
-        {
-            buf = [0u8; 16];
-            let (prefix, r) =
-                Ref::<_, [u8]>::unaligned_from_suffix_with_elems(&mut buf[..], 8).unwrap();
-            assert_eq!(prefix, [0; 8]);
-            test_new_helper_slice_unaligned(r, 8);
-        }
-    }
-
-    #[test]
-    fn test_new_oversized() {
-        // Test that a properly-aligned, overly-sized buffer works for
-        // `new_from_prefix` and `new_from_suffix`, and that they return the
-        // remainder and prefix of the slice respectively.
-
-        let mut buf = Align::<[u8; 16], AU64>::default();
-        {
-            // In a block so that `r` and `suffix` don't live too long. `buf.t`
-            // should be aligned to 8, so this should always succeed.
-            let (r, suffix) = Ref::<_, AU64>::from_prefix(&mut buf.t[..]).unwrap();
-            assert_eq!(suffix.len(), 8);
-            test_new_helper(r);
-        }
-        {
-            buf.set_default();
-            // `buf.t` should be aligned to 8, so this should always succeed.
-            let (prefix, r) = Ref::<_, AU64>::from_suffix(&mut buf.t[..]).unwrap();
-            assert_eq!(prefix.len(), 8);
-            test_new_helper(r);
-        }
-    }
-
-    #[test]
-    fn test_new_unaligned_oversized() {
-        // Test than an unaligned, overly-sized buffer works for
-        // `new_unaligned_from_prefix` and `new_unaligned_from_suffix`, and that
-        // they return the remainder and prefix of the slice respectively.
-
-        let mut buf = [0u8; 16];
-        {
-            // In a block so that `r` and `suffix` don't live too long.
-            let (r, suffix) = Ref::<_, [u8; 8]>::unaligned_from_prefix(&mut buf[..]).unwrap();
-            assert_eq!(suffix.len(), 8);
-            test_new_helper_unaligned(r);
-        }
-        {
-            buf = [0u8; 16];
-            let (prefix, r) = Ref::<_, [u8; 8]>::unaligned_from_suffix(&mut buf[..]).unwrap();
-            assert_eq!(prefix.len(), 8);
-            test_new_helper_unaligned(r);
-        }
-    }
-
-    #[test]
-    fn test_ref_from_mut_from() {
-        // Test `FromBytes::{ref_from, mut_from}{,_prefix,Suffix}` success cases
-        // Exhaustive coverage for these methods is covered by the `Ref` tests above,
-        // which these helper methods defer to.
-
-        let mut buf =
-            Align::<[u8; 16], AU64>::new([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]);
-
-        assert_eq!(
-            AU64::ref_from(&buf.t[8..]).unwrap().0.to_ne_bytes(),
-            [8, 9, 10, 11, 12, 13, 14, 15]
-        );
-        let suffix = AU64::mut_from(&mut buf.t[8..]).unwrap();
-        suffix.0 = 0x0101010101010101;
-        // The `[u8:9]` is a non-half size of the full buffer, which would catch
-        // `from_prefix` having the same implementation as `from_suffix` (issues #506, #511).
-        assert_eq!(
-            <[u8; 9]>::ref_from_suffix(&buf.t[..]).unwrap(),
-            (&[0, 1, 2, 3, 4, 5, 6][..], &[7u8, 1, 1, 1, 1, 1, 1, 1, 1])
-        );
-        let (prefix, suffix) = AU64::mut_from_suffix(&mut buf.t[1..]).unwrap();
-        assert_eq!(prefix, &mut [1u8, 2, 3, 4, 5, 6, 7][..]);
-        suffix.0 = 0x0202020202020202;
-        let (prefix, suffix) = <[u8; 10]>::mut_from_suffix(&mut buf.t[..]).unwrap();
-        assert_eq!(prefix, &mut [0u8, 1, 2, 3, 4, 5][..]);
-        suffix[0] = 42;
-        assert_eq!(
-            <[u8; 9]>::ref_from_prefix(&buf.t[..]).unwrap(),
-            (&[0u8, 1, 2, 3, 4, 5, 42, 7, 2], &[2u8, 2, 2, 2, 2, 2, 2][..])
-        );
-        <[u8; 2]>::mut_from_prefix(&mut buf.t[..]).unwrap().0[1] = 30;
-        assert_eq!(buf.t, [0, 30, 2, 3, 4, 5, 42, 7, 2, 2, 2, 2, 2, 2, 2, 2]);
-    }
-
-    #[test]
     fn test_ref_from_mut_from_error() {
         // Test `FromBytes::{ref_from, mut_from}{,_prefix,Suffix}` error cases.
 
@@ -7710,79 +6594,6 @@ mod tests {
         assert!(AU64::mut_from_prefix(&mut buf.t[1..]).is_err());
         assert!(AU64::ref_from_suffix(&buf.t[..]).is_err());
         assert!(AU64::mut_from_suffix(&mut buf.t[..]).is_err());
-    }
-
-    #[test]
-    #[allow(clippy::cognitive_complexity)]
-    fn test_new_error() {
-        // Fail because the buffer is too large.
-
-        // A buffer with an alignment of 8.
-        let buf = Align::<[u8; 16], AU64>::default();
-        // `buf.t` should be aligned to 8, so only the length check should fail.
-        assert!(Ref::<_, AU64>::from(&buf.t[..]).is_err());
-        assert!(Ref::<_, [u8; 8]>::unaligned_from(&buf.t[..]).is_err());
-
-        // Fail because the buffer is too small.
-
-        // A buffer with an alignment of 8.
-        let buf = Align::<[u8; 4], AU64>::default();
-        // `buf.t` should be aligned to 8, so only the length check should fail.
-        assert!(Ref::<_, AU64>::from(&buf.t[..]).is_err());
-        assert!(Ref::<_, [u8; 8]>::unaligned_from(&buf.t[..]).is_err());
-        assert!(Ref::<_, AU64>::from_prefix(&buf.t[..]).is_err());
-        assert!(Ref::<_, AU64>::from_suffix(&buf.t[..]).is_err());
-        assert!(Ref::<_, [u8; 8]>::unaligned_from_prefix(&buf.t[..]).is_err());
-        assert!(Ref::<_, [u8; 8]>::unaligned_from_suffix(&buf.t[..]).is_err());
-
-        // Fail because the length is not a multiple of the element size.
-
-        let buf = Align::<[u8; 12], AU64>::default();
-        // `buf.t` has length 12, but element size is 8.
-        assert!(Ref::<_, [AU64]>::from(&buf.t[..]).is_err());
-        assert!(Ref::<_, [[u8; 8]]>::unaligned_from(&buf.t[..]).is_err());
-
-        // Fail because the buffer is too short.
-        let buf = Align::<[u8; 12], AU64>::default();
-        // `buf.t` has length 12, but the element size is 8 (and we're expecting
-        // two of them).
-        assert!(Ref::<_, [AU64]>::from_prefix_with_elems(&buf.t[..], 2).is_err());
-        assert!(Ref::<_, [AU64]>::from_suffix_with_elems(&buf.t[..], 2).is_err());
-        assert!(Ref::<_, [[u8; 8]]>::unaligned_from_prefix_with_elems(&buf.t[..], 2).is_err());
-        assert!(Ref::<_, [[u8; 8]]>::unaligned_from_suffix_with_elems(&buf.t[..], 2).is_err());
-
-        // Fail because the alignment is insufficient.
-
-        // A buffer with an alignment of 8. An odd buffer size is chosen so that
-        // the last byte of the buffer has odd alignment.
-        let buf = Align::<[u8; 13], AU64>::default();
-        // Slicing from 1, we get a buffer with size 12 (so the length check
-        // should succeed) but an alignment of only 1, which is insufficient.
-        assert!(Ref::<_, AU64>::from(&buf.t[1..]).is_err());
-        assert!(Ref::<_, AU64>::from_prefix(&buf.t[1..]).is_err());
-        assert!(Ref::<_, [AU64]>::from(&buf.t[1..]).is_err());
-        assert!(Ref::<_, [AU64]>::from_prefix_with_elems(&buf.t[1..], 1).is_err());
-        assert!(Ref::<_, [AU64]>::from_suffix_with_elems(&buf.t[1..], 1).is_err());
-        // Slicing is unnecessary here because `new_from_suffix` uses the suffix
-        // of the slice, which has odd alignment.
-        assert!(Ref::<_, AU64>::from_suffix(&buf.t[..]).is_err());
-
-        // Fail due to arithmetic overflow.
-
-        let buf = Align::<[u8; 16], AU64>::default();
-        let unreasonable_len = usize::MAX / mem::size_of::<AU64>() + 1;
-        assert!(Ref::<_, [AU64]>::from_prefix_with_elems(&buf.t[..], unreasonable_len).is_err());
-        assert!(Ref::<_, [AU64]>::from_suffix_with_elems(&buf.t[..], unreasonable_len).is_err());
-        assert!(Ref::<_, [[u8; 8]]>::unaligned_from_prefix_with_elems(
-            &buf.t[..],
-            unreasonable_len
-        )
-        .is_err());
-        assert!(Ref::<_, [[u8; 8]]>::unaligned_from_suffix_with_elems(
-            &buf.t[..],
-            unreasonable_len
-        )
-        .is_err());
     }
 
     #[test]
@@ -7890,45 +6701,6 @@ mod tests {
         let foo = Foo { a: [0xFFFF; 33] };
         let expected = [0xFFu8; 66];
         assert_eq!(foo.as_bytes(), &expected[..]);
-    }
-
-    #[test]
-    fn test_display_debug() {
-        let buf = Align::<[u8; 8], u64>::default();
-        let r = Ref::<_, u64>::from(&buf.t[..]).unwrap();
-        assert_eq!(format!("{}", r), "0");
-        assert_eq!(format!("{:?}", r), "Ref(0)");
-
-        let buf = Align::<[u8; 8], u64>::default();
-        let r = Ref::<_, [u64]>::from(&buf.t[..]).unwrap();
-        assert_eq!(format!("{:?}", r), "Ref([0])");
-    }
-
-    #[test]
-    fn test_eq() {
-        let buf1 = 0_u64;
-        let r1 = Ref::<_, u64>::from(buf1.as_bytes()).unwrap();
-        let buf2 = 0_u64;
-        let r2 = Ref::<_, u64>::from(buf2.as_bytes()).unwrap();
-        assert_eq!(r1, r2);
-    }
-
-    #[test]
-    fn test_ne() {
-        let buf1 = 0_u64;
-        let r1 = Ref::<_, u64>::from(buf1.as_bytes()).unwrap();
-        let buf2 = 1_u64;
-        let r2 = Ref::<_, u64>::from(buf2.as_bytes()).unwrap();
-        assert_ne!(r1, r2);
-    }
-
-    #[test]
-    fn test_ord() {
-        let buf1 = 0_u64;
-        let r1 = Ref::<_, u64>::from(buf1.as_bytes()).unwrap();
-        let buf2 = 1_u64;
-        let r2 = Ref::<_, u64>::from(buf2.as_bytes()).unwrap();
-        assert!(r1 < r2);
     }
 
     #[test]
